app/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain_openai
openai
chromadb
pypdf
requests
pydantic-settings
python-dotenv
fitz
pymupdf